+++ client/src/TableSensors.js
@@ -228,8 +228,7 @@ class TableSensors extends React.Component {
 
 
     handleChange(name, value) {
-        if (isNumber(parseInt(value))) { var val = parseInt(value) } else { var val = value };
-
+        if (isNumber(parseFloat(value))) { var val = parseFloat(value) } else { var val = value };
         this.setState({ [name]: val });
     };
